+++ src/main/scala/edu/ie3/simona/agent/participant2/ParticipantAgentInit.scala
@@ -8,6 +8,7 @@ package edu.ie3.simona.agent.participant2
 
 import edu.ie3.datamodel.models.input.system.SystemParticipantInput
 import edu.ie3.simona.agent.grid.GridAgent
+import edu.ie3.simona.agent.participant.statedata.ParticipantStateData.InputModelContainer
 import edu.ie3.simona.agent.participant2.ParticipantAgent._
 import edu.ie3.simona.config.RuntimeConfig.BaseRuntimeConfig
 import edu.ie3.simona.event.ResultEvent
@@ -83,9 +84,9 @@ object ParticipantAgentInit {
 
   /** Starts the initialization process of a [[ParticipantAgent]].
     *
-    * @param participantInput
-    *   The system participant model input that represents the physical model at
-    *   the core of the agent.
+    * @param inputContainer
+    *   The input container holding the system participant model input that
+    *   represents the physical model at the core of the agent.
     * @param runtimeConfig
     *   Runtime configuration that has to match the participant type.
     * @param notifierConfig
@@ -101,7 +102,7 @@ object ParticipantAgentInit {
     *   [[edu.ie3.simona.agent.em.EmAgent]].
     */
   def apply(
-      participantInput: SystemParticipantInput,
+      inputContainer: InputModelContainer[_ <: SystemParticipantInput],
       runtimeConfig: BaseRuntimeConfig,
       notifierConfig: NotifierConfig,
       participantRefs: ParticipantRefs,
@@ -114,11 +115,11 @@ object ParticipantAgentInit {
 
         em ! RegisterControlledAsset(
           flexAdapter,
-          participantInput,
+          inputContainer.electricalInputModel,
         )
 
         em ! ScheduleFlexActivation(
-          participantInput.getUuid,
+          inputContainer.electricalInputModel.getUuid,
           INIT_SIM_TICK,
         )
 
@@ -141,7 +142,7 @@ object ParticipantAgentInit {
       }
 
     uninitialized(
-      participantInput,
+      inputContainer,
       runtimeConfig,
       notifierConfig,
       participantRefs,
@@ -153,7 +154,7 @@ object ParticipantAgentInit {
   /** Waiting for an [[Activation]] message to start the initialization.
     */
   private def uninitialized(
-      participantInput: SystemParticipantInput,
+      inputContainer: InputModelContainer[_ <: SystemParticipantInput],
       runtimeConfig: BaseRuntimeConfig,
       notifierConfig: NotifierConfig,
       participantRefs: ParticipantRefs,
@@ -166,11 +167,11 @@ object ParticipantAgentInit {
       // first, check whether we're just supposed to replay primary data time series
       participantRefs.primaryServiceProxy ! PrimaryServiceRegistrationMessage(
         ctx.self.toClassic,
-        participantInput.getUuid,
+        inputContainer.electricalInputModel.getUuid,
       )
 
       waitingForPrimaryProxy(
-        participantInput,
+        inputContainer,
         runtimeConfig,
         notifierConfig,
         participantRefs,
@@ -184,7 +185,7 @@ object ParticipantAgentInit {
     * participant uses model calculations or just replays primary data.
     */
   private def waitingForPrimaryProxy(
-      participantInput: SystemParticipantInput,
+      inputContainer: InputModelContainer[_ <: SystemParticipantInput],
       runtimeConfig: BaseRuntimeConfig,
       notifierConfig: NotifierConfig,
       participantRefs: ParticipantRefs,
@@ -205,7 +206,7 @@ object ParticipantAgentInit {
 
       completeInitialization(
         ParticipantModelShell.createForPrimaryData(
-          participantInput,
+          inputContainer,
           runtimeConfig,
           primaryDataExtra,
           simulationParams.simulationStart,
@@ -221,7 +222,7 @@ object ParticipantAgentInit {
     case (ctx, RegistrationFailedMessage(_)) =>
       // we're _not_ supposed to replay primary data, thus initialize the physical model
       val modelShell = ParticipantModelShell.createForPhysicalModel(
-        participantInput,
+        inputContainer,
         runtimeConfig,
         simulationParams.simulationStart,
         simulationParams.simulationEnd,
@@ -254,7 +255,7 @@ object ParticipantAgentInit {
 
         requiredServices.foreach { case (serviceType, serviceRef) =>
           registerForService(
-            participantInput,
+            inputContainer.electricalInputModel,
             ctx.self,
             modelShell,
             serviceType,