+++ src/test/scala/edu/ie3/simona/agent/participant2/ParticipantAgentInitSpec.scala
@@ -9,6 +9,7 @@ package edu.ie3.simona.agent.participant2
 import edu.ie3.datamodel.models.OperationTime
 import edu.ie3.simona.agent.grid.GridAgent
 import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ActivePowerExtra
+import edu.ie3.simona.agent.participant.statedata.ParticipantStateData.SimpleInputContainer
 import edu.ie3.simona.agent.participant2.ParticipantAgent.{
   PrimaryRegistrationSuccessfulMessage,
   RegistrationFailedMessage,
@@ -71,13 +72,15 @@ class ParticipantAgentInitSpec
 
     val operationStart = 10 * 3600L
 
-    val mockInput = loadInput.copy
-      .operationTime(
-        OperationTime.builder
-          .withStart(operationStart.toDateTime)
-          .build()
-      )
-      .build()
+    val mockInput = SimpleInputContainer(
+      loadInput.copy
+        .operationTime(
+          OperationTime.builder
+            .withStart(operationStart.toDateTime)
+            .build()
+        )
+        .build()
+    )
 
     "not controlled by EM" should {
 
@@ -116,7 +119,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -163,7 +166,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -208,12 +211,15 @@ class ParticipantAgentInitSpec
         )
 
         val emRegistrationMsg = em.expectMessageType[RegisterControlledAsset]
-        emRegistrationMsg.modelUuid shouldBe mockInput.getUuid
-        emRegistrationMsg.inputModel shouldBe mockInput
+        emRegistrationMsg.modelUuid shouldBe mockInput.electricalInputModel.getUuid
+        emRegistrationMsg.inputModel shouldBe mockInput.electricalInputModel
         val activationRef = emRegistrationMsg.participant
 
         em.expectMessage(
-          ScheduleFlexActivation(mockInput.getUuid, INIT_SIM_TICK)
+          ScheduleFlexActivation(
+            mockInput.electricalInputModel.getUuid,
+            INIT_SIM_TICK,
+          )
         )
 
         activationRef ! FlexActivation(INIT_SIM_TICK)
@@ -221,7 +227,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -231,7 +237,7 @@ class ParticipantAgentInitSpec
 
         em.expectMessage(
           FlexCompletion(
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
             requestAtTick = Some(operationStart),
           )
         )
@@ -265,12 +271,15 @@ class ParticipantAgentInitSpec
         )
 
         val emRegistrationMsg = em.expectMessageType[RegisterControlledAsset]
-        emRegistrationMsg.modelUuid shouldBe mockInput.getUuid
-        emRegistrationMsg.inputModel shouldBe mockInput
+        emRegistrationMsg.modelUuid shouldBe mockInput.electricalInputModel.getUuid
+        emRegistrationMsg.inputModel shouldBe mockInput.electricalInputModel
         val activationRef = emRegistrationMsg.participant
 
         em.expectMessage(
-          ScheduleFlexActivation(mockInput.getUuid, INIT_SIM_TICK)
+          ScheduleFlexActivation(
+            mockInput.electricalInputModel.getUuid,
+            INIT_SIM_TICK,
+          )
         )
 
         activationRef ! FlexActivation(INIT_SIM_TICK)
@@ -278,7 +287,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -289,7 +298,10 @@ class ParticipantAgentInitSpec
         )
 
         em.expectMessage(
-          FlexCompletion(mockInput.getUuid, requestAtTick = Some(15 * 3600L))
+          FlexCompletion(
+            mockInput.electricalInputModel.getUuid,
+            requestAtTick = Some(15 * 3600L),
+          )
         )
       }
 
@@ -301,13 +313,15 @@ class ParticipantAgentInitSpec
 
     val operationStart = 10 * 3600L
 
-    val mockInput = pvInput.copy
-      .operationTime(
-        OperationTime.builder
-          .withStart(operationStart.toDateTime)
-          .build()
-      )
-      .build()
+    val mockInput = SimpleInputContainer(
+      pvInput.copy
+        .operationTime(
+          OperationTime.builder
+            .withStart(operationStart.toDateTime)
+            .build()
+        )
+        .build()
+    )
 
     val runtimeConfig = PvRuntimeConfig()
 
@@ -349,7 +363,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -360,8 +374,8 @@ class ParticipantAgentInitSpec
         service.expectMessage(
           RegisterForWeatherMessage(
             participantAgent.toClassic,
-            mockInput.getNode.getGeoPosition.getY,
-            mockInput.getNode.getGeoPosition.getX,
+            mockInput.electricalInputModel.getNode.getGeoPosition.getY,
+            mockInput.electricalInputModel.getNode.getGeoPosition.getX,
           )
         )
 
@@ -410,7 +424,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -460,12 +474,15 @@ class ParticipantAgentInitSpec
         )
 
         val emRegistrationMsg = em.expectMessageType[RegisterControlledAsset]
-        emRegistrationMsg.modelUuid shouldBe mockInput.getUuid
-        emRegistrationMsg.inputModel shouldBe mockInput
+        emRegistrationMsg.modelUuid shouldBe mockInput.electricalInputModel.getUuid
+        emRegistrationMsg.inputModel shouldBe mockInput.electricalInputModel
         val activationRef = emRegistrationMsg.participant
 
         em.expectMessage(
-          ScheduleFlexActivation(mockInput.getUuid, INIT_SIM_TICK)
+          ScheduleFlexActivation(
+            mockInput.electricalInputModel.getUuid,
+            INIT_SIM_TICK,
+          )
         )
 
         activationRef ! FlexActivation(INIT_SIM_TICK)
@@ -473,7 +490,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -484,8 +501,8 @@ class ParticipantAgentInitSpec
         service.expectMessage(
           RegisterForWeatherMessage(
             participantAgent.toClassic,
-            mockInput.getNode.getGeoPosition.getY,
-            mockInput.getNode.getGeoPosition.getX,
+            mockInput.electricalInputModel.getNode.getGeoPosition.getY,
+            mockInput.electricalInputModel.getNode.getGeoPosition.getX,
           )
         )
 
@@ -495,7 +512,10 @@ class ParticipantAgentInitSpec
         )
 
         em.expectMessage(
-          FlexCompletion(mockInput.getUuid, requestAtTick = Some(12 * 3600L))
+          FlexCompletion(
+            mockInput.electricalInputModel.getUuid,
+            requestAtTick = Some(12 * 3600L),
+          )
         )
       }
 
@@ -527,12 +547,15 @@ class ParticipantAgentInitSpec
         )
 
         val emRegistrationMsg = em.expectMessageType[RegisterControlledAsset]
-        emRegistrationMsg.modelUuid shouldBe mockInput.getUuid
-        emRegistrationMsg.inputModel shouldBe mockInput
+        emRegistrationMsg.modelUuid shouldBe mockInput.electricalInputModel.getUuid
+        emRegistrationMsg.inputModel shouldBe mockInput.electricalInputModel
         val activationRef = emRegistrationMsg.participant
 
         em.expectMessage(
-          ScheduleFlexActivation(mockInput.getUuid, INIT_SIM_TICK)
+          ScheduleFlexActivation(
+            mockInput.electricalInputModel.getUuid,
+            INIT_SIM_TICK,
+          )
         )
 
         activationRef ! FlexActivation(INIT_SIM_TICK)
@@ -540,7 +563,7 @@ class ParticipantAgentInitSpec
         primaryService.expectMessage(
           PrimaryServiceRegistrationMessage(
             participantAgent.ref.toClassic,
-            mockInput.getUuid,
+            mockInput.electricalInputModel.getUuid,
           )
         )
 
@@ -553,7 +576,10 @@ class ParticipantAgentInitSpec
         )
 
         em.expectMessage(
-          FlexCompletion(mockInput.getUuid, requestAtTick = Some(15 * 3600L))
+          FlexCompletion(
+            mockInput.electricalInputModel.getUuid,
+            requestAtTick = Some(15 * 3600L),
+          )
         )
 
         // service should not be called at all