+++ src/test/scala/edu/ie3/simona/test/common/input/EvcsInputTestData.scala
@@ -12,6 +12,7 @@ import edu.ie3.datamodel.models.input.system.EvcsInput
 import edu.ie3.datamodel.models.input.system.`type`.chargingpoint.ChargingPointTypeUtils
 import edu.ie3.datamodel.models.input.system.`type`.evcslocation.EvcsLocationType
 import edu.ie3.datamodel.models.input.system.characteristic.CosPhiFixed
+import edu.ie3.simona.agent.participant.statedata.ParticipantStateData.SimpleInputContainer
 import edu.ie3.simona.test.common.DefaultTestData
 import edu.ie3.simona.test.common.model.MockEvModel
 import edu.ie3.util.quantities.QuantityUtils.RichQuantityDouble
@@ -34,6 +35,8 @@ trait EvcsInputTestData extends DefaultTestData with NodeInputTestData {
     EvcsLocationType.HOME,
     true,
   )
+
+  protected val evcsInputContainer = SimpleInputContainer(evcsInputModel)
 
   protected val ev1 = new MockEvModel(
     UUID.fromString("0-0-0-1-1"),